backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
httpx
mongoengine
pymongo
motor
coqui-tts
mongoengine
